This commit includes the implementation of a web-based data import mechanism.
When data is imported, the user may choose:
- optin with email confirmation request
- optout with email notification
- optout with no notification
The resulting e-mail will be sent to the primary e-mail address on record for an organization, if there is one.
Confirmation e-mails include links to APPROVE the entry (via a unique token) or to become the EDITOR of the entry (user must register/log in).
Notification e-mails include links to EDIT the entry or become an EDITOR (both require login), or to RESTRICT or HIDE the entry (via a unique token).
Both e-mail types include a link to display the current entry as it (appears/would appear) on the website in an IFRAME in the e-mail. I originally tried rendering the display directly in the e-mail but it became complicated & problematic for various reasons.

Other updates:
- Admin menu was moved to the left side of the screen, and is now vertical.
- DSO menu appears in the same part of the screen, under Admin menu if both are visible, or instead of if only one.